requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-otp==1.5.4
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
